server/requirements.txt
mem0ai>=0.1.0
flask>=3.0.0
chromadb>=0.4.0
openai>=1.0.0
